[api] Handle maintenance statistics requests for remotes
Forward maintenance statistics requests to remote instance and respond
with the received xml. If the response is empty, the remote project didn't
exist and we respond with a 404.
We don't do a redirect here, since we want to be consistent with the
other places where we have to handle remote instances.