[webui] Fix build log for multibuild and local link packages
The @package variable was set in the check_build_log_access before filter.
However, for multibuild and local link packages this was set to the parent package (follow_mulitbuild / follow_project_links)
which does not contain a log file. This resulted that set_initial_offset returned an empty log and calculated no offset
which eventually resulted in cutting the end of the log files.
Fix #2651.