[ci] Use last Rantly version instead patched gem
Use last Rantly version instead our own patched gem, as the last version
includes our patch.

In Travis we were using Rantly 1.0.0 instead the patched version. So the
Rantly output could not be disabled in this case as there was important
information not rendered in the error message. Get disabled now.