[ci] Fix failing revision tests
Some of the revision tests where depending on a package that was also
used by other tests. Since one of the newly added tests were also
caused a revision change, some of the revision tests started to fail.

This commit removes the package dependency and refactors parts of the
revision tests:

* Moved creation of revisions to package factory
* Made tests independent from package objects used by other tests
* Added description to some of the 'it' blocks to make it clearer what
  is actually meant to be tested there
* Updated comments for backend cleanup