[ci] Use a separate package factory for trigger_service tests.
Some of our tests rely on data stored in the backend. Since factory girl
doesn't provide an 'after destroy' hook, we either have to manually clean
this up, or use separate object for tests that rely on backend data.

Given that only a few tests seem to rely on backend data we go for the
latter.

This was found after deleting extisting VCR cassettes and re-running the
tests.