[api] TEMPORARY FIX Show packages with dot in name
If we have for a example a package called `ana.yaml`,
`/package/show/home:Admin/ana.yaml` will be processed as YAML althought
the `.yaml` is part of the package name and it is not intended to be
the request format.

This seems to be a Rails bug, this is just a temporarly fix. Other
routes are also affected, but these is the more visible for user,
so I think that at least here it is worthwhile to provide a
temporary fix.

Fixes https://github.com/openSUSE/open-build-service/issues/2900