[api] OBSEngine::Base#mount_it is not used
We never create an instance of `OBSEngine::Base`, so having and
instance method is pointless.

What we actually do is calling the class method with the same name in
all the subclasses. So, the `mount_it` method in `OBSEngine::Base`, but
I guess the idea was to leave the method also in the parent class as
help. In this case, this method should be the class method and not the
instance one.