[api][webui] Add polymorphic association to review
entities are user, group, project and package.
This makes ActiveModel queries more simple and efficient.
Also introduced validation that setting only one entity is allowed.
Furthermore changed the collation of the by_ fields to utf8_general_ci
as joining on these fields was not possible (e.g. group#title had
a mismatched collation).
This was already done on build.o.o manually in the past so
this commit will make the structures more consistent.