[webui] Set the default format of package routes to html
Fixes #2900

Since Rails 5 the routing changed to support the new API application
feature in Rails. For OBS that meant that packages with a mime-like
suffix, like .yaml, caused Rails to handle the response as "mime-type"
format, eg. yaml, instead of html.

To fix this we have to explicitly set the default format to html to
route properly.

Comment from the rails guys:
  https://github.com/rails/rails/issues/28901#event-1060620644