[ci] Disable haml-linters with many offenses
In the autogenerated Haml-Lint todo we used to list as excluded all the
files with offenses. This looked horrible as for some linters we had
many affected files.

Now when we have more than 15 files we exclude the lint instead,
similarly as it is done in Rubocop and as it will be done by Haml-Lint
soon.