[webui] Fix creating a git service via webui
Creating a git service by adding a git repo url in the package add file
view stopped working. This was because URI#parse is not able to parse
git ssh urls.
In case of http urls to github repositories, OBS was creating a download
file service instead of the expected scm service.

Additional change: Second parameter of addService is now an empty array
instead of nil. That's because addService would fail for nil values

Both will be fixed by this commit.