[backend] add and make use of new get_path_projpacks method
The problem we need to solve is that with native docker builds,
we get repositories comming with the container. That is, our path
setup depends on a expandsion result. Unfortunately this means
that our projpacks and remotemap data may be incomplete, so we
need to do a second fetch to fill the missing data.

We'll also need to do this in the scheduler, but this will be
more tricky as we do not want to do synchronous requests.