[webui] Bugfix: Change error message to notice for spiders
With commit 357fd7cddc040cb6527ddd20cc4a67ae8cdf3f3c I introduced a
linter error (guard clause was not used) and a error in our spider test
case because we showed them an error flash if the _reason file of a
package is empty. This gets fixed with that commit. We're now using a
notice since it's actually a notice and not an error.