[webui] Use less javascript/css for codemirror
We really do not need to include over 150 javascript files supporting
a gazillion file types when we know which file type we are going to
use in the editor.

The sourcediff partial also loaded all of this code twice! I'm also
pretty sure that all of this was making the feature tests we run
for requests flicker.

If you feel adventurous, webpack this! (After we upgrade to RoR 5.1)