[app][webui] generalize rescue in update_from_xml
The `update_from_xml` method is only used from
ProjectController#save_meta and
ConsistencyCheckJob#import_project_from_backend.
In the first one we want to get the error instead of rescuing the
`APIException` to show the error to the UI. In the second case,
we are just ignoring this Exception. So I changed `update_from_xml` by
`update_from_xml!` in this case and rescue the `APIException` there.
It is missed to decide if we want to keep ignoring this exception
or not, so I added a FIXME message there.

Fixes https://github.com/openSUSE/open-build-service/issues/3140