[ci] Fix Kiwi Repository model Rspec test
We were using `@` as an invalid character, although it is not. I
changed it by `{`. We were also not checking all the protocols that
would make the test not fail.

I also add more explicit test descriptions and make more restrictive
the test cases, to ensure that we test the invalid case that we want
and that it doesn't fail just by coincidence.

Fixes https://github.com/openSUSE/open-build-service/issues/3247