[ci] Precompile assets when running rspec tests in travis
When travis is on higher load it can happen that the first rspec test
using poltergeist is failing with a "Capybara::Poltergeist::StatusFailError:
check DNS and/or server status" error.

The poltergeist maintainers recommend[*] to precompile the assets to
lower the likelyness for this to happen. This commit is doing so...

[*] https://github.com/teampoltergeist/poltergeist/issues/791