[ci] Add a linter for the structure.sql
In our rpm package builds we run a set of migrations, dump the
db structure afterwards and compare it with the one stored in git.

If they differ, the package builds fail.

Now it can happen that the structure.sql is not touched, but gems that
affect how the dumped strucutre.sql looks like changed and result in a
different format, eg. using different chars to mark keywords, adding the
indexing method even when it's the default one, etc. .

To avoid this from happening we add a new linter that runs in travis.