[backend] make Docker building work for partitions
This is not an easy task. The problem is that for Docker builds,
the container provides extra path elements. Thus we do not know
the complete path when getting the project data (getprojpack).
This results in missing entries in our remotemap.

This commit tries to deal with that problem by requesting remotemap
updates from the build job generation code. We also need a new
remotemissing hash so that we do not request non-existing projects
over and over again.

For now we change the info element from the package data by
appending the new path elements. This has the advantage that
the project sorting function will see the new dependencies and
do the right thing. Patching the info data is a bit nasty, though.