[webui] Split reviewable polymorphic association into single associations
We need to have seperate associations for project and package because a Project maintainer should see Reviews by_package.
This was not easily possible with the polymorphic association as the review only contained the package_id and not the project_id in this case
(only with extensive joining this would have been possible).
Fixes #3282.