[backend] put base container annotation in the build job
We do this so that the containerinfo generation code can get the
repository data of the base image.

This is obviously a hack. The correct way would be to introspect
the docker image. We hopefully will do that in the future.

(But note that the repositories in the image are unordered,
while we currently need ordered repos.)