[ci] Fix flickering test: backend_build_command
Fix flickering test: `Package#backend_build_command`.

In most tests for package model we had this:

```
before do
  login(user)
end
```

where `user.login` is `tom`

And there is one test which had instead:

```
let(:other_user) { create(:user) }

before do
  login(other_user)
end
```

So in this case `user.login` is not fixed.

For the `Package#backend_build_command` we didn't login any user, so
there were three options:

- The user was not logged in -> `nobody_user`
- `tom` is logged in -> `tom` (most of the cases)
- `other_user` is logged in -> not fixed name

So, the cassets were generating with `tom` as user, so in the other
two cases this fail. You can see an example of failure here:

github.com/openSUSE/open-build-service/pull/3300#issuecomment-312222370

To fix it I just log `tom` for this tests.
I also remove `let(:other_user) { create(:user) }` as there is up
already an `other_user` and with a fixed name`. I regenerate
all the cassettes for Package and check that `other_user` is only there
for the test using it.