[ci] Implement script to download and import a db dump
This script downloads database dump from a remote server and imports it into your local database.
It is a script and not a rake task, because downloading from a remote server usually needs
a private ssh key. As we're using vagrant in development this would mean that we need to
copy the private key into the vagrant box which we avoid with this script.