[api] hotfix maintenance incident processing
Fixes a crash when packages gets referenced mutiple times

It might be better to touch the project in Package.store, when
it creates a new instance. But I wanted to have something as small
as possible due to time pressure.

It might be better to touch the project in Package.store, when
it creates a new instance. But I wanted to have something as small
as possible due to time pressure.