[backend] add instantiate flag to sourcecopy
This can be used to instantiate a package coming from a project
link. We need this because we first have to write the new meta file
to create the package container, but after that the package will
be gone. So the new flag makes the sourcecopy ignore the local
package and instead search the project link.