[webui] Remove action count from page title
This information is not actionable from the page title.
The user cannot perform the actions when activating the tab.
Users frequently have a gazillion of obs tabs open, it just does not look pretty.
The count is repeated in the page header text where it IS actionable by pulling up the action list.
The count is frequently outdated and not refreshed.

This partially reverts commit 41e5d2e093fbc0a0eaf7ea6b1581166568e818b8.
Fixes #3407