[api] ldap: Simplify checking of Config values
This shortens the code for checking if a value for a certain
configuration option matches what we inspect.
When a key value pair does not exist in the Config store, the return
value for that non-existing key is nil. Therefore we can skip checking
the existence of a key and just query andcompare it in one go.