[ci] Update test expectation of request controller test
Changes expected state of request from 'revoked' to 'declined' as well
as the expected comment.

Package#close_requests closes requests as
  - declined when the target package gets deleted
  - revoked when the source package gets deleted

Because of that the correct expectation for the test should be that the
request is in state 'declined' and the comment tag contains a hint that
the target package was deleted.