[webui] Add `external` column to `users` table
It's not really clear if a user comes from extern or signed up locally.
This can harm us if we want to distinguish between users who signed up
locally (for example the Admin user) and people that come from LDAP or
some other service. For this, we need an extra column in the `users`
table. The field `adminnote` is not really a good idea to match against
it since it's a string and can also be used for different stuff in the
future.

Users which signed up externally have the field `adminnote` set to `User
created via LDAP`. We can safely migrate all existing users since the
field `adminnote` is not yet changable without directly manipulating the
field in the database. @adrianschroeter said it's a safe approach.