[dist] Generate assets in %post
Rails expects a clean public/assets directory. It does not expect files in
there that assets:precompile does not generate, which can easily be the case if
an package update goes wrong or someone fiddled with the directory in some
other way. Only generating the assets during package buid did not make sure
that this is covered.

By generating them during %install AND in %post we make sure that asset
compilation works and that the directory is clean.