[ci][webui] Fix handling global roles in user edit
Fixes #3566

The user save action is handling editing user accounts by users that want
to update their account as well as admins that edit other user's
accounts or their own.
In the latter case the code was accidentally dropping all global roles
of the admin user.

In addition there was a mismatch between the model implementation (allows
multiple roles) and the views (only showed the Admin role).
Since this changed recently with the introduction of the 'staff' role,
we had to fix this.
We now show all available global roles in the admin user edit page.

Or in short:

* Fixes dropping of global roles when an admin edits his own account (#3566)
* Allow admins to see and edit all global roles available for a user
* Improve tests for save action