[api][webui] Fix sending broken event over and over again
When an event from the backend is broken (e.g. invalid character like emjoi in commit message),
this job tries to create the event over and over again in an infinite loop.
The problem was that the rescue was outside of the loop.
This commit moved the rescue again inside the loop and partially reverts commit
2bee297917a01ece4f00804a1b581b9e04e563f9.