[api] add logging to try and debug IssueTracker#parse_github_issue
Sometimes it gets a response from github which it cannot parse but we
dont know what that response is or why it gets it so I've added logging
to try and figure out whats going on. The previous Airbrake logging
wasn't working because the extra data (json_data) was not being sent to
errbit because Airbrake.notify does not allow custom data to be sent.