[ci] Fix flickering failure of this spec.
The problem was that all the contexts in this spec used the same
project & package and they also rely on the test backend running.
So once one context ran it meant the backend package existed in the
backend for a project with that name already for the other contexts.
This cause failures when running without the test backend using the
record VCR cassettes.