[frontend] Fix drop tags rollback
Tags where dropped in
https://github.com/openSUSE/open-build-service/pull/3644

The problem is that from Rails 5.1 the default primary keys are
BIGINT(20). That means that the references is trying to use BIGINT(20)
while the old tables use INT(11). So we have to take this into account,
otherwise the rollback will fail.

I also removed some duplicated indexes, as over some field we were
creating an index and an unique index.

Mob programmed by @bgeuken and @Ana06.