[api] Hide sourcediffs in event mails in case of remote projects
Fixes #3587

When generating a sourcediff OBS checks the source access. In case of
remote projects this access check includes checking the write access of
the logged in user. Since event mails are sent by delayed jobs, there is no
user available to run the source access check against, which causes the
delayed job to crash.

A proper would require a bigger refactoring of the code that is involved
here. For the time being we work around the issue by skipping the source
diff generation for event emails that involve remote projects.