[ci] Drop code quality test
This test model contains two tests. A static ruby syntax check and
a code complexity test.

The static code analyzation is covered by rubocop, and our test suite.
For measuring code complexity we already use codeclimate in OBS.

Beside this both tests don't seem to fullfil any purpose. We know that
we have very complex and ugly code in OBS. And whenever we touch it we
try to improve things, if time permits.

For the cases where time does not permit, we are forced to update the
complexity test. In other words we ignore it.

And if we actiuvely want to check out OBS code complexity, or even
improve it, we are still able to fetch these data from
  https://codeclimate.com/github/openSUSE/open-build-service