[api][webui] Fix Event creation when comment is too long.
The max char length of comment.body is 65535 but the max char length of
events.payload is also 65535. Since the payload contains the comment
body and other data, its possible that the comment body will not fit in
the event payload. This commit adds a callback which shortens the
comment body if it is necessary to fit it in the payload.