[api] Handle database data changes with a gem.
Use the data-migrate gem to handle changes to data with their own set of
migrations just like we handle changes to the database structure.

This gives us a set structure to update existing data in the database
and gives us a single command to run (rake db:migrate:with_data) in
deployments.