[api] avoid an endless loop in special situations
When multiple reuse of repositories in same project in the right order.
Current code will find the new created path again and again, so
we need a cycle detection here.

It would be possible to simplify the code a lot, if we just could update
a single entry instead, but we don't guarantee single entry for position
atm.