[api] improve database operation in project repository path relink handling
We do still see many dead locks in production since more projects triggering
this code exist.

Problem is that the index grouping of INNODB leads to dead locks when
multiple delete/create operations are running (also on indepedend entries).

So, let's optimize for the situation that really only a single matching entry exists
and do an UPDATE operation in that case. Avoiding the need to touch the indexes.