[api][webui] Convert bigint columns to int & foreign keys.
Since rails 5.1 changed the default primary key type from int to
bigint we now either have to convert all primary keys and foreign keys
to bigint (which will take a lot of time) or stick with type int. So
this commit changes the 4 instances of bigint to int so we will stick with int for the meantime.