[dist][api] Update handling of user-agents config
We used to continuously fetch the crawler-user-agents.json from
upstream. This recently started to cause build failures, because a new
Phantomjs bot got add to the config.
Since our test suite is using phantomsjs and we run the test suite
during the package build, this broke our builds.

This commit solves this by
  * Stopping to fetch the config continuously and instead add the
    crawler-user-agents.json file into our git repo
  * Dropping the phantomjs entry from the configuration