[ci] Fix Rubocop
We need to set the Ruby target version in Rubocop to 2.4 due to a bug
in the parser that makes our test suite fail. Changing the Ruby version
introduced new offenses that had been automatically added to the todo
file. After that, a `Lint/UnneededDisable` offense appeared, due to
having added that file to the todo file. It was automatically
corrected.