[webui] Remove unused exception handling
UnknownObjectError is not raised by any of the methods called in the
controller. This exception can occure only when calling update_from_xml.
But this method is wrapped within a transaction, which will catch the
exception.

This commit removes the unused rescue and adds a test for calling
save_meta with invalid devel projects.
We also update the xml parser code to throw a more explicit error
message.