[ci] Solve duplicated shared exampls warning
It would also be an option to merge both examples. But that would mean
we have to deal with 2 variables that can change (subject and review /
request). This would increase the complexity of the example a lot.
For that reason it's better to just fix the name collision.

====

WARNING: Shared example group 'the subject's cache is reset' has been previously defined at:
  /obs/src/api/spec/models/bs_request_spec.rb:119
...and you are now defining it at:
  /obs/src/api/spec/models/review_spec.rb:321
The new definition will overwrite the original one.