[webui] Disable links for remote image templates
We were linking to projects and packages of remote instances. In such a
case we don't have a valid url to point at. So users would end up with
an error when clicking this link.

To fix that we only show a link for local projects and packages on the
image template page.