[frontend] Create history element on priority raise
Part of the bs request validation is to check, if the priority of a
associated bs_request_action is higher than the priority of the
bs_request.
If that's the case, the priority of the bs_request get's raised.

This wasn't logged via the history elements and might be one of the
culprits of Issue#3897 – history elements were missing.