[backend] [api] check the sha256 of file if needed
This will set $entry->{'hash'} = 'missing' if there is already a
file on the server with the same md5sum, but in another project
with the same package name.

The client(osc) will then calculate the sha256 for the file and return
it to the srcserver.

The sha256 sum of the uploaded file will be compared to the one from
the one already existing on the server.

added sha256 check in BSSrcrep::addfile