[api][webui] Set default format of comments route to xml
The /comments/ route is also accessible via the webui. Since we don't
have a html representation for that resource, we would end with an empty
page.

This commits sets the default format of that route to xml. So that users
see the xml representation instead of a blank page.

Fixes #3180