[ci] Regenerate Rubocop todo file
When moving the code into different classes the exclusion in the
`rubocop_todo.yml` doesn't have effect any more.

Also, we had an inclussion of a helper outside a class/module. Rubocop
was not complainning before due to a bug, that is already fixed in
master and will be in the next release. Regenerating the todo file solve
this as well. As we need to solve this in more places, it seems to be
fine to exclude it by now.