[api] Remove Event::Package
`Event::Package` is a class with almost nothing that is used as
superclass for:

- `Event::BranchCommand`
- `Event::Build`
- `Event::CommentForPackage`
- `Event::Commit`
- `Event::CreatePackage`
- `Event::DeletePackage`
- `Event::ServiceFail`
- `Event::ServiceSuccess`
- `Event::UndeletePackage`
- `Event::UpdatePackage`
- `Event::Upload`
- `Event::VersionChange`

The `Event::Package` causes problem as we also have a `Package` model
in the root level. So once one of the two packages models is loaded,
Rails doens't look for the second one, which lead to uses of the wrong
model. To avoid that Rails uses the `Package` in the root level in
`Event::Package` subclasses we can call the Package specifying the
module. But for the other case there is nothing we can do. This is all
caused because we shouldn't have models with repeated names even if
they are in different levels. As the `Event::Package` is useless, we can
just remove it and add its data directly to the subclasses.