[webui][api] Fix CleanupEvents job
Before PR #4197 was merged we would set project_logged = true for all events older
than 10 days. This no longer happens. So we need to make sure we delete those events
who are not of the type which has a ProjectLogEntry created, regardless of whether
or not event.project_logged is set to true or false.