[webui][api] Fix CreateProjectLogEntryJob when the entry was not persisted.
We have ~2300 events in the db which cannot have a ProjectLogEntry created
for them because the project they relate to no longer exists and project_id
is a required field on ProjectLogEntry. In the past these events would have
been left there and deleted after 10 days but this commit makes it so they get
deleted at the end of the day.