[backend] cloudupload: support getting multiple jobs and the log file
Log file support is a bit nasty because we want to use the AJAX
server for this so that we don't use up server slots. But
streaming of plain files was not implemented yet, as the workers
just use a server process and rely on the 'joinable' code in
the source/repo server.