[frontend] Update html selector in flags test
Flags are shown in a table filled with icons. When hovering over an
item of that table a menu appears.
The test was previously using the text field of the invisible menu.
Since this is not reflecting what a user would do, we use the visible
icon instead. That icon is matched via the '.current_flag_state' html
class.

As a side effect we also dropped the now unused 'from' parameter.