[frontend] Fix redirect of GroupsController#update
While group maintainers are allowed to edit a group they maintain they
are not permitted to see all existing groups.
So when a group maintainer edited a group successfully, he still saw a
flash message telling him ''.

Changing the redirect to the edit page instead fixes the error. This
also means that the user sees the group he just edited instead of all
existing groups.