[frontend] Refactor Re-assignment of controller parameters
Uses slice to reduce the code for re-assigning key/value pairs to a new
hash.

Because the branch action is not permitting parameters, we have to
convert the parameters to a hash first.

Also not that fetching the key / value pairs does no longer rely on the
parameter values, but the existance of the key. This is not an issue,
since these parameters are either intentionally set or don't exist.