[frontend] Downgrade bunny to version 2.8.0
With the last deployment sending to rabbitmq broke and caused errors
("No CA certificates found, add one with :tls_ca_certificates").

We checked the certificate setup together with Adrian and they seem to
be fine. Given that, and that the issue started with the last
deployment, it seems likely that the recent bunndy update to version
2.9.0 caused the issue.

That's why we downgrade the package for now until we found the root
cause.

Mob-programming with @DavidKang, @mdeniz