[backend] Link the file into the working directory
and get rid of the temporary hack --target-filename.
All files transfered to the uploader are called ID.file, this was a problem for ec2uploadimg as
it was not able to determine the file extension.
Now we link the ID.file into the working directory with the actual filename and upload this file.