[frontend] Fix default subscription switch
PR#4336 added a switch that allows users to reset their notification
subscription to default. This was hardcoded to clear all subscriptions.

With this commit the default subscription settings are retrieved from
the server via ajax and properly set in the ui.
The user can then decide to save the new settings or cancel it by
reloading the page.