[frontend] Replace routing spec with controller test
A rubocop-rspec cop was complaining about an offense that could not be
corrected, and was not added to the rubocop todo file.

There were a couple of issues with the spec (long version below), which
caused the test to never run in our test suite.
I decided to replace the routing test, because the feature switch we
were testing is defined on controller level. So with a routing test it
wouldn't be possible to touch / test this code.

The long version:

There were a couple of issues with the spec:

* The spec was not having a _spec.rb suffix and thus never run in our
  test suite
* Fixing the wrong file name caused test failures because of calling
  the controller in a non working way (expect(get: :image_templates))
* Fixing this would still cause test failures because of the usages of
  Rails.application.reloader.reload!