[dist] Add a .dockerignore file
We are using our main directory as our so called docker build context.
That means that when doing a build all files and directories of our
repository where sent to the docker daemon.

Since we only need some of the files, we can reduce this quite a lot by
adding unneeded files and dirs to a docker ignore file.

This reduced the build context size from 557.8MB to 171.9MB on my
machine (with 220MB of log files). I dumped my development.log before,
which had a size of ~4GB ;-)

The build time (real time) went down from 24s to 12s (both using cached
images).