[doc] Introduces a commit message template
A bunch of us have talked about what the content of good commit meesages should
look like and came up with this template. Here is how you can make use of it:

cp .gitmessage ~/.gitmessage

Add this to your ~/.gitconfig

[commit]
  template = ~/.gitmessage

Everyboy following this template will help people to understand the changes in
the future. Addtionally it will make it easier for us to use commit messages in
documentation like our Changelog, books or our blog.

Co-authored-by: Manuel Schnitzer <github@mschnitzer.de>
Co-authored-by: Björn Geuken <bgeuken@suse.de>
Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>
Co-authored-by: Moises Deniz Aleman <mdeniz@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>