[dist] Enable cloud upload feature on backend
This enables the cloud upload feature in the backend of our docker
environment.

It also adds a new public key located in /etc/obs/cloudupload/_pubkey
which will be served via the API:
http://$srcserver:$port/cloudupload/_pubkey

The secret key is in the keyring:
    gpg --armor --export-secret-keys obs-cloud-uploader-dev@opensuse.org

The secret key will later be used in the uploading script for decrypting
sensitive data such as the Azure management certificate which comes
encrypted from the webui.