[frontend] Fix ec2 upload errors
The vpc_subnet_id parameter was not removed from the parameter list
before doing the request to the backend. Because the backend validates
these parameters and doesn't expect / allow a vpc_subnet_id it caused a
failure during the upload.

Disabled rubocop rule, since this test is just about sending the correct
request. Having this rule enabled would mean that we have to configure
the backend for cloud uploads.