[frontend] A default for Configuration.obs_url
Configuration.obs_url is needed for any email sending to work and it's not
obviouos that this is the case. Neither is it explained in our documentation or
setup scripts. So let's set some fake default from here on out.

We don't migrate any existing data because this would have the side effect that
all of the sudden emails would be sent, which we would like to avoid.