[webui] increase contrast on default text
Owing to the physics of monitors, there is almost no visible
difference between 0x44 gray and actual black. (The lowest gray you
can find in the classic X11 color list even is 0x69, which is
probably reflecting just that fact.)

A block of 0x44 text is perceived as black, and only when 0x00 text
is right next to it, a "boldening" effect on the 0x00 portion is
noticable, at which point it feels like a weird design choice to
mix that "semi-bold" and "non-bold" text.