[webui][ci] Fix test for Packages controller
Due to we removed the `required_parameter` method and add the parameters in the routes
we need to pass those missing parameters in the redirect_to callback.

We also removed a test that checked for an 404 that is not needed. That
is already handled by the routes.

Add cassettes for the modified tests.

Co-authored-by: David Kang <dkang@suse.com>