[webui] Remove required_parameters from add_reviewer
Remove `required_parameters` callback in `add_reviewer`.
As `review_type` comes from fixed values in a `select_tag` we can rely
on the code of the `rescue` clauses to deal with values not included in
the `select_tag`.

This will allow us to remove the following class in the future:
``class MissingParameterError < RuntimeError; end``.

Co-authored-by: David Kang <dkang@suse.com>