[backend] rework watch setup
We no longer create watches in the middle of our project dependency
calculation. Instead we do it after we're done with recalculation.
It's much nicer and less error prone this way. To do this change,
we had to introduce a new hash 'expandedprojlink' that contains
the expanded project link data.