[webui] Fix encodings exception in event mailer
When creating a mail with a diff between two packages, this diff can be
encoded in a different encoding as UTF-8. In this case, mixing this `diff`
field in the view used by the mailer with other fields like `description`
of the request that are in UTF-8 results in throwing this exception:
"incompatible character encodings".

Making sure that the `diff` field is converted to UTF-8 in the view used by the
mailer prevents this exception to occur.

Test added for this case.

Fixes #3415.

Co-authored-by: Ana María Martínez Gómez <ammartinez@suse.de>
Co-authored-by: David Kang <dkang@suse.com>
Co-authored-by: Moisés Déniz Alemán <mdeniz@suse.com>