[frontend] Remove redundant assignment
If the "package_change?" method is called and a project parameter is
passed, it seems to be a String (I only checked the "obvious" callers
by grepping for "package_change?"). Hence, there is no need to do
any "conversion" etc. Originally, it was introduced in commit
f2b734be14 ("backward compatibility to handle also string arguments").