[webui] Rescue `mail` callbacks in `EventMailer`
As there is no information regarding the parameters that cause an
`ArgumentError` exception in the `mail` callback in `EventMailer`, this
exception is catched. It will only log the parameters to the error log,
and then will re-raise the exception.

This will help to know the reasons `ArgumentError` exception is thrown
and therefore help to fix #4052.

Co-authored-by: David Kang <dkang@suse.com>