[frontend] Merge SetBugowner und AddRole requests
`BsRequestActionSetBugowner` is just the same as
`BsRequestActionAddRole`, with the difference that in the first one
previous bugowners are removed. For this small difference we are
duplicating plenty of code.

I would even suggest to get rid of `BsRequestActionSetBugowner`, but
let's start with an small change.

Apart from the main difference already mentioned, there were some others
between the code of these two classes:

- `Project.find_by_name!(target_project)` in SetBugowner vs
  `Project.find_by_name(target_project)` in AddRole.
  But I guess we want it to always fail if project is not found.
  So this can be unified.

- We set `ignore_lock` to add the role in SetBugowner. This come from:
  https://github.com/openSUSE/open-build-service/commit/cb15ab2b4dd981e23da922327a77cd53aca41402